+++ src/egon/data/datasets/scenario_parameters/parameters.py
@@ -142,9 +142,7 @@ def electricity(scenario):
 
         costs = read_csv(2035)
 
-        parameters = {
-            "grid_topology": "Status Quo",
-        }
+        parameters = {"grid_topology": "Status Quo"}
         # Insert effciencies in p.u.
         parameters["efficiency"] = {
             "oil": read_costs(costs, "oil", "efficiency"),
@@ -303,7 +301,31 @@ def electricity(scenario):
         }
 
     elif scenario == "eGon100RE":
+        costs = read_csv(2050)
         parameters = {"grid_topology": "Status Quo"}
+        # Insert effciencies in p.u.
+        parameters["efficiency"] = {
+            "battery": {
+                "store": read_costs(costs, "battery inverter", "efficiency")
+                ** 0.5,
+                "dispatch": read_costs(costs, "battery inverter", "efficiency")
+                ** 0.5,
+                "standing_loss": 0,
+                "max_hours": 6,
+            },
+            "pumped_hydro": {
+                "store": read_costs(costs, "PHS", "efficiency") ** 0.5,
+                "dispatch": read_costs(costs, "PHS", "efficiency") ** 0.5,
+                "standing_loss": 0,
+                "max_hours": 6,
+            },
+        }
+        # Insert capital costs
+        parameters["capital_cost"] = {
+            "battery": read_costs(costs, "battery inverter", "investment")
+            + parameters["efficiency"]["battery"]["max_hours"]
+            * read_costs(costs, "battery storage", "investment")  # [EUR/MW]
+        }
 
     else:
         print(f"Scenario name {scenario} is not valid.")
@@ -419,10 +441,18 @@ def gas(scenario):
             "CH4_to_H2": read_costs(costs, "SMR", "investment"), # CC?
             "H2_feedin": 0,
             "H2_to_CH4": read_costs(costs, "methanation", "investment"),
-            "H2_underground": read_costs(costs, "hydrogen storage underground", "investment"),
-            "H2_overground": read_costs(costs, "hydrogen storage tank incl. compressor", "investment"),
-            "H2_pipeline": read_costs(costs, "H2 (g) pipeline", "investment"),  # [EUR/MW/km]
-            "H2_pipeline_retrofit": read_costs(costs, "H2 (g) pipeline repurposed", "investment"),  # [EUR/MW/km]
+            "H2_underground": read_costs(
+                costs, "hydrogen storage underground", "investment"
+            ),
+            "H2_overground": read_costs(
+                costs, "hydrogen storage tank incl. compressor", "investment"
+            ),
+            "H2_pipeline": read_costs(
+                costs, "H2 (g) pipeline", "investment"
+            ),  # [EUR/MW/km]
+            "H2_pipeline_retrofit": read_costs(
+                costs, "H2 (g) pipeline repurposed", "investment"
+            ),  # [EUR/MW/km]
         }
         parameters["marginal_cost"] = {
             "CH4": global_settings(scenario)["fuel_costs"]["gas"]
@@ -430,6 +460,9 @@ def gas(scenario):
             * global_settings(scenario)["co2_emissions"]["gas"],
             "OCGT": read_costs(costs, "OCGT", "VOM"),
         }
+
+    elif scenario == "eGon100RE":
+        parameters = {"main_gas_carrier": "H2"}
 
     else:
         print(f"Scenario name {scenario} is not valid.")